Add actual coverability-check (optional) (#9)
* Move property checking into own file.

* Small refactorings.

* Implement basic & slow coverability

* Add TODOs

* Use i16 instead of i32

* Update frontend on unbounded RG response.

* Move coverability to own module.

* Remove global step check.

* determine strictly_greater_than in one iteration and move TODOs

* Small refactoring.

* Display RG info message.

* Improve layout of info-bar

+++ src-tauri/src/common.rs
@@ -2,11 +2,11 @@ use derive_new::new;
 use ndarray::{Array1, Array2, Axis, s};
 use serde::Serialize;
 
-pub(crate) fn fire_transition(state: &Array1<i32>, effect_matrix: &Array2<i32>, t: usize) -> Array1<i32> {
+pub(crate) fn fire_transition(state: &Array1<i16>, effect_matrix: &Array2<i16>, t: usize) -> Array1<i16> {
     state + &effect_matrix.slice(s![t, ..])
 }
 
-pub(crate) fn find_active_transitions(marking: &Array1<i32>, transition_inputs: &Array2<i32>) -> Vec<i32> {
+pub(crate) fn find_active_transitions(marking: &Array1<i16>, transition_inputs: &Array2<i16>) -> Vec<i16> {
     let mut active_transitions = Vec::new();
 
     // Compare each row of the matrix to the reference array
@@ -14,14 +14,14 @@ pub(crate) fn find_active_transitions(marking: &Array1<i32>, transition_inputs: 
         // Check whether the marking is at least as large as the edge weight.
         if marking.iter().zip(row.iter()).all(|(&a, &b)| a >= b) {
             // If the marking is large enough, the transition is active.
-            active_transitions.push(row_index as i32);
+            active_transitions.push(row_index as i16);
         }
     }
 
     return active_transitions;
 }
 
-pub(crate) fn vec_vec_to_array2(input: &Vec<Vec<i32>>, rows: &usize, columns: &usize) -> Array2<i32> {
+pub(crate) fn vec_vec_to_array2(input: &Vec<Vec<i16>>, rows: &usize, columns: &usize) -> Array2<i16> {
     let mut result = Array2::zeros((*rows, *columns));
     for (i, mut row) in result.axis_iter_mut(Axis(0)).enumerate() {
         for (j, col) in row.iter_mut().enumerate() {
@@ -34,8 +34,8 @@ pub(crate) fn vec_vec_to_array2(input: &Vec<Vec<i32>>, rows: &usize, columns: &u
 
 #[derive(Serialize, new)]
 pub struct SimulationResponse {
-    marking: Vec<i32>,
-    firings: Vec<i32>,
+    marking: Vec<i16>,
+    firings: Vec<i16>,
     deadlocked: bool,
 }
